Move gql models
